refactor: move duplicate methods to common interfaces
- Added ConditionStep and LogicalStep common interfaces to keep
  duplicate methods
- Protected type safety with generics
