+++ src/main/java/org/example/finlog/query_builder/builder/SelectQueryBuilder.java
@@ -3,8 +3,8 @@ package org.example.finlog.query_builder.builder;
 import org.example.finlog.query_builder.step.general.BuildStep;
 import org.example.finlog.query_builder.step.select.*;
 import org.example.finlog.query_builder.step.where.ComparisonStep;
-import org.example.finlog.query_builder.step.where.ConditionStepForSelect;
-import org.example.finlog.query_builder.step.where.LogicalStepForSelect;
+import org.example.finlog.query_builder.step.where.condition.ConditionStepForSelect;
+import org.example.finlog.query_builder.step.where.logical.LogicalStepForSelect;
 import org.example.finlog.query_builder.util.Operation;
 import org.example.finlog.query_builder.util.OrderDirection;
 import org.example.finlog.query_builder.util.QuerySerializer;
