+++ src/main/java/org/example/finlog/query_builder/builder/UpdateQueryBuilder.java
@@ -6,8 +6,8 @@ import org.example.finlog.query_builder.step.general.ValueStep;
 import org.example.finlog.query_builder.step.update.SetStep;
 import org.example.finlog.query_builder.step.update.UpdateStep;
 import org.example.finlog.query_builder.step.where.ComparisonStep;
-import org.example.finlog.query_builder.step.where.ConditionStepForUpdate;
-import org.example.finlog.query_builder.step.where.LogicalStepForUpdate;
+import org.example.finlog.query_builder.step.where.condition.ConditionStepForUpdate;
+import org.example.finlog.query_builder.step.where.logical.LogicalStepForUpdate;
 import org.example.finlog.query_builder.util.Operation;
 import org.example.finlog.query_builder.util.QuerySerializer;
 import org.example.finlog.query_builder.util.SqlKeyWord;
