+++ src/main/java/org/example/finlog/query_builder/step/select/FromStep.java
@@ -1,6 +1,6 @@
 package org.example.finlog.query_builder.step.select;
 
-import org.example.finlog.query_builder.step.where.ConditionStepForSelect;
+import org.example.finlog.query_builder.step.where.condition.ConditionStepForSelect;
 
 public interface FromStep {
     ConditionStepForSelect from(String table);
