--- src/main/java/org/example/finlog/query_builder/step/where/ConditionStepForUpdate.java
@@ -1,7 +1,0 @@
-package org.example.finlog.query_builder.step.where;
-
-import org.example.finlog.query_builder.step.general.BuildStep;
-
-public interface ConditionStepForUpdate extends BuildStep {
-    ComparisonStep<LogicalStepForUpdate> where(String field);
-}
