+++ src/main/java/org/example/finlog/query_builder/step/where/condition/ConditionStep.java
@@ -1,0 +1,8 @@
+package org.example.finlog.query_builder.step.where.condition;
+
+import org.example.finlog.query_builder.step.where.ComparisonStep;
+import org.example.finlog.query_builder.step.where.logical.LogicalStep;
+
+public interface ConditionStep<NEXT extends LogicalStep<NEXT>> {
+    ComparisonStep<NEXT> where(String field);
+}
