+++ src/main/java/org/example/finlog/query_builder/step/where/condition/ConditionStepForSelect.java
@@ -1,0 +1,13 @@
+package org.example.finlog.query_builder.step.where.condition;
+
+import org.example.finlog.query_builder.step.general.BuildStep;
+import org.example.finlog.query_builder.step.select.LimitStep;
+import org.example.finlog.query_builder.step.select.OrderStep;
+import org.example.finlog.query_builder.step.where.logical.LogicalStepForSelect;
+
+public interface ConditionStepForSelect extends
+        ConditionStep<LogicalStepForSelect>,
+        OrderStep,
+        LimitStep,
+        BuildStep {
+}
