+++ src/main/java/org/example/finlog/query_builder/step/where/logical/LogicalStepForUpdate.java
@@ -1,0 +1,8 @@
+package org.example.finlog.query_builder.step.where.logical;
+
+import org.example.finlog.query_builder.step.general.BuildStep;
+
+public interface LogicalStepForUpdate extends
+        LogicalStep<LogicalStepForUpdate>,
+        BuildStep {
+}
